requirements.txt
streamlit>=1.34.0
pandas>=2.0.0
numpy>=1.23.0
yfinance>=0.2.36
wikipedia-api>=0.6.0
feedparser>=6.0.10
nltk>=3.8.1
sentence-transformers>=3.0.0
scikit-learn>=1.3.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dateutil>=2.8.2
